fix: First attempt at refining Personal Savings Allowance calculation
